requirements.txt
asgiref==3.2.10
cffi==1.14.2
cryptography==3.1
Django==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
psycopg2==2.8.5
pycparser==2.20
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
